added merge sort algorithm

+++ Main.cpp
@@ -10,7 +10,7 @@ using namespace std;
 
 int main(){
 
-    List<int> list(0,2,4,1,3,5,6,64,4);
+    List<float> list(0,2,4,1,3,5,6,64,4);
     list.sort();
     list.print();
